devops/roles/icos.typesense/templates/requirements.txt
beautifulsoup4==4.13.1
typesense==1.0.3
requests==2.32.3
pyyaml==6.0.2
ftfy==6.3.1
